llama-index-packs/llama-index-packs-code-hierarchy/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.codespell]
check-filenames = true
check-hidden = true
skip = "*.csv,*.html,*.json,*.jsonl,*.pdf,*.txt,*.ipynb"

[tool.llamahub]
contains_example = false
import_path = "llama_index.packs.code_hierarchy"

[tool.llamahub.class_authors]
CodeHierarchyAgentPack = "ryanpeach"

[tool.mypy]
disallow_untyped_defs = true
exclude = ["_static", "build", "examples", "notebooks", "venv"]
ignore_missing_imports = true
python_version = "3.8"

[tool.poetry]
authors = ["Ryan Peach <rgpeach10@gmail.com>"]
description = "A node parser which can create a hierarchy of all code scopes in a directory."
keywords = ["c", "c_sharp", "code", "cpp", "elisp", "elm", "go", "hierarchy", "java", "javascript", "ocaml", "php", "python", "ql", "ruby", "rust", "tree-sitter", "tree-sitter-languages", "typescript"]
license = "MIT"
maintainers = ["ryanpeach"]
name = "llama-index-packs-code-hierarchy"
readme = "README.md"
version = "0.2.0"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
llama-index-core = "^0.10.1"
tree-sitter-languages = "^1.8.0"
tree-sitter = "^0.20.2"
llama-index-agent-openai = "^0.1.5"
llama-index-readers-file = "^0.1.8"
gitpython = "^3.1.42"

[tool.poetry.group.dev.dependencies]
ipython = "8.10.0"
jupyter = "^1.0.0"
pytest = "7.2.1"

[tool.poetry.group.dev.dependencies.black]
extras = ["jupyter"]
version = "<=23.9.1,>=23.7.0"

[tool.poetry.group.dev.dependencies.codespell]
extras = ["toml"]
version = ">=v2.2.6"

[[tool.poetry.packages]]
include = "llama_index/"
